moving server and client under a 'cmd' folder
also changing the proto & code that uses the generated code that results
