+++ Makefile
@@ -10,16 +10,11 @@ DEV_ENV_CMD := ${DEV_ENV_PREFIX} ${DEV_ENV_IMAGE}
 BUILD_ALPINE_CMD_PREFIX := ${DEV_ENV_PREFIX} -e CGO_ENABLED=0 -e GOOS=linux -e GOARCH=amd64 ${DEV_ENV_IMAGE}
 
 bootstrap:
-	${DEV_ENV_CMD} glide install
+	glide install
 
-build:
-	go build -o goprox
-
-build-alpine:
-	${BUILD_ALPINE_CMD_PREFIX} go build -o rootfs/bin/goprox
-
-build-cli:
-	make -C cli build
+build: 
+	make -C cmd/server build
+	make -C cmd/cli build
 
 docker-build:
 	docker build --rm -t ${DOCKER_IMAGE_NAME} rootfs
@@ -27,16 +22,8 @@ docker-build:
 docker-push:
 	docker push ${DOCKER_IMAGE_NAME}
 
-docker-deploy: build-alpine docker-build docker-push
-
 test:
-	${DEV_ENV_CMD} go test $$(glide nv)
+	go test $$(glide nv)
 
-run-test:
-	docker run --rm -e AWS_KEY=${GOPROX_AWS_KEY} -e AWS_SECRET=${GOPROX_AWS_SECRET} ${DOCKER_IMAGE_NAME}
-
-codegen-admin:
-	make -C ./_proto codegen-admin
-
-deploy-to-deis:
-	${DEIS_BINARY_NAME} pull ${DOCKER_IMAGE_NAME} -a ${DEIS_APP_NAME}
+codegen:
+	make -C ./_proto codegen
