--- _proto/admin.proto
@@ -1,24 +1,0 @@
-syntax = "proto3";
-
-package admin;
-
-service Admin {
-  rpc GetPackages(Empty) returns (PackageList) {}
-  rpc UpgradePackage(FullPackage) returns (FullPackage) {}
-  rpc AddPackage(FullPackage) returns (FullPackage) {}
-}
-
-message FullPackage {
-  string Name = 1;
-  string SHA = 2;
-}
-
-message PackageList {
-  repeated Package packages = 1;
-}
-
-message Package {
-  string Name = 1;
-}
-
-message Empty {}
