--- cli/Makefile
@@ -1,2 +1,0 @@
-build:
-	go build -o goprox
