+++ cmd/cli/Makefile
@@ -1,0 +1,2 @@
+build:
+	go build -o goprox .
