+++ cmd/cli/main.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/codegangsta/cli"
-	"google.golang.org/grpc"
+	// "google.golang.org/grpc"
 )
 
 const (
