+++ cmd/server/Makefile
@@ -1,0 +1,2 @@
+build:
+	go build -o goproxd
